Support transparency in track color
Instead of drawing read ahead and progress on top of the segments with gaps with SrcIn blend mode, the segments are always drawn without gaps, then read ahead and progress overlaid with the default SrcOver blend mode, and only then gaps are removed via Clear blend mode. See #12 for more details on the original problem.